Part a updated.

+++ src/Triangle.java
@@ -9,7 +9,7 @@ public class Triangle extends Shape{
     double getArea() {
         double p;
         double area;
-        p = (a + b + c) / 2;
+        p = (double)(a + b + c) / 2;
         area = Math.sqrt(p * (p - a) * (p - b) * (p - c));
         return area;
     }
